% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\title{Indices}
\usage{
indices(len, len.out)
}
\arguments{
\item{len}{The length of the vector to be symbolized.}

\item{len.out}{The length of x after dimensional reduction.}
}
\value{
The indices of PAA.
}
\description{
Indices
}
\examples{
indices(100, 10)
}
