% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paa_indices.R
\name{paa_indices}
\alias{paa_indices}
\title{PAA Indices}
\usage{
paa_indices(len, len.out)
}
\arguments{
\item{len}{The length of the vector to be symbolized.}

\item{len.out}{The length of x after dimensional reduction.}
}
\value{
The indices for doing Piecewise Aggregate Approximation.
}
\description{
PAA Indices
}
\examples{
paa_indices(100, 10)
}
