% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remainder_increment.R
\name{remainder_increment}
\alias{remainder_increment}
\title{Increment Remainder}
\usage{
remainder_increment(len, len.out)
}
\arguments{
\item{len}{The length of the original vector.}

\item{len.out}{The length of the vector after reduction.}
}
\value{
A vector with increments.
}
\description{
Increment Remainder
}
\examples{
remainder_increment(1000, 100)
}
