% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssymbolize.R
\name{ssymbolize}
\alias{ssymbolize}
\title{SSymbolize}
\usage{
ssymbolize(x, alpha, len.out, type, from, to)
}
\arguments{
\item{x}{The vector to be symbolized.}

\item{alpha}{The size of the alphabet.}

\item{len.out}{The output length.}

\item{type}{The SAX method.}

\item{from}{The beginning of the desired interval.}

\item{to}{The end of the desired interval.}
}
\value{
A symbolic vector.
}
\description{
SSymbolize
}
