% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolize.R
\name{symbolize}
\alias{symbolize}
\title{Symbolize}
\usage{
symbolize(x, alpha, w, type, alphabet = letters)
}
\arguments{
\item{x}{The vector to be symbolized.}

\item{alpha}{The size of the alphabet.}

\item{w}{The output length.}

\item{type}{The SAX method.}

\item{alphabet}{The alphabet to be used.}
}
\value{
A symbolic vector.
}
\description{
Symbolize
}
