% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{indices}
\alias{indices}
\title{Title}
\usage{
indices(x, len.out)
}
\arguments{
\item{x}{The vector to be symbolized.}

\item{len.out}{The length of x after dimensionality reduction.}
}
\value{
The indices ...
}
\description{
Title
}
\examples{
indices(rnorm(100), 10)
}
